added profile route where it shows you the faveourites of user logged in, the username and also a button to sign out
